refactor(editor): Move remaining canvas code to features (no-changelog)
