--- packages/frontend/editor-ui/src/__tests__/data/index.ts
@@ -1,1 +1,0 @@
-export * from './canvas';
