+++ packages/frontend/editor-ui/src/features/canvas/__tests__/utils.ts
