+++ packages/frontend/editor-ui/src/features/canvas/canvas.eventBus.ts
@@ -1,4 +1,4 @@
-import type { CanvasEventBusEvents } from '@/features/canvas/canvas.types';
+import type { CanvasEventBusEvents } from './canvas.types';
 import { createEventBus } from '@n8n/utils/event-bus';
 
 export const canvasEventBus = createEventBus<CanvasEventBusEvents>();
