+++ packages/frontend/editor-ui/src/features/canvas/components/elements/handles/render-types/CanvasHandleNonMainInput.test.ts
@@ -1,6 +1,6 @@
 import CanvasHandleNonMainInput from './CanvasHandleNonMainInput.vue';
 import { createComponentRenderer } from '@/__tests__/render';
-import { createCanvasHandleProvide } from '@/__tests__/data';
+import { createCanvasHandleProvide } from '@/features/canvas/__tests__/utils';
 
 const renderComponent = createComponentRenderer(CanvasHandleNonMainInput);
