+++ packages/frontend/editor-ui/src/features/canvas/components/elements/handles/render-types/parts/CanvasHandlePlus.test.ts
@@ -1,7 +1,7 @@
 import { fireEvent } from '@testing-library/vue';
 import CanvasHandlePlus from './CanvasHandlePlus.vue';
 import { createComponentRenderer } from '@/__tests__/render';
-import { createCanvasHandleProvide } from '@/__tests__/data';
+import { createCanvasHandleProvide } from '@/features/canvas/__tests__/utils';
 
 const renderComponent = createComponentRenderer(CanvasHandlePlus, {
 	global: {
