+++ packages/frontend/editor-ui/src/features/canvas/components/elements/nodes/CanvasNodeRenderer.test.ts
@@ -1,6 +1,6 @@
 import CanvasNodeRenderer from './CanvasNodeRenderer.vue';
 import { createComponentRenderer } from '@/__tests__/render';
-import { createCanvasNodeProvide, createCanvasProvide } from '@/__tests__/data';
+import { createCanvasNodeProvide, createCanvasProvide } from '@/features/canvas/__tests__/utils';
 import { createTestingPinia } from '@pinia/testing';
 import { setActivePinia } from 'pinia';
 import { CanvasNodeRenderType } from '../../../canvas.types';
