+++ packages/frontend/editor-ui/src/features/canvas/components/elements/nodes/render-types/CanvasNodeDefault.test.ts
@@ -1,7 +1,7 @@
 import CanvasNodeDefault from './CanvasNodeDefault.vue';
 import { createComponentRenderer } from '@/__tests__/render';
 import { NodeConnectionTypes } from 'n8n-workflow';
-import { createCanvasNodeProvide, createCanvasProvide } from '@/__tests__/data';
+import { createCanvasNodeProvide, createCanvasProvide } from '@/features/canvas/__tests__/utils';
 import { createTestingPinia } from '@pinia/testing';
 import { setActivePinia } from 'pinia';
 import { CanvasConnectionMode, CanvasNodeRenderType } from '../../../../canvas.types';
