+++ packages/frontend/editor-ui/src/features/logs/composables/useLogsSelection.ts
@@ -7,7 +7,7 @@ import {
 	isSubNodeLog,
 } from '@/features/logs/logs.utils';
 import { useTelemetry } from '@/composables/useTelemetry';
-import { canvasEventBus } from '@/event-bus/canvas';
+import { canvasEventBus } from '@/features/canvas/canvas.eventBus';
 import type { IExecutionResponse } from '@/Interface';
 import { useCanvasStore } from '@/stores/canvas.store';
 import { useLogsStore } from '@/stores/logs.store';
