+++ packages/frontend/editor-ui/src/features/ui/commandBar/composables/useNodeCommands.ts
@@ -5,7 +5,7 @@ import { useCredentialsStore } from '@/stores/credentials.store';
 import { useRootStore } from '@n8n/stores/useRootStore';
 import { useCanvasOperations } from '@/composables/useCanvasOperations';
 import { useActionsGenerator } from '@/components/Node/NodeCreator/composables/useActionsGeneration';
-import { canvasEventBus } from '@/event-bus/canvas';
+import { canvasEventBus } from '@/features/canvas/canvas.eventBus';
 import { type CommandBarItem } from '@n8n/design-system/components/N8nCommandBar/types';
 import { getNodeIcon, getNodeIconUrl } from '@/utils/nodeIcon';
 import type { SimplifiedNodeType } from '@/Interface';
